WIP - Scaled Health/Damage by 10 for skills dealing minimal amounts of damage a lot of times. Added Weapon Mace: Basic Swing ignores a portion of the targets armor. Second Skill reduces the targets armor for a duration (allies can benefit from this). Added Weapon: Arcane Staff: Energy Ray: A Ray that is active as long as the button is pressed. Deals damage and pulls enemies closer. Energy Bubble: A bubble that is active as long as the button is pressed. Increases armor for allies inside it. Added/Changed some skill functionalities to allow these skills.

+++ Assets/Scripts/Combat/Test/SkillEffects/SkillEffectInflictCondition.cs
@@ -8,15 +8,15 @@ public class SkillEffectInflictCondition : SkillEffect {
     [Header("Skill Effect Inflict Condition:")]
     public Condition ConditionToApply;
 
-    public int DurationBase = 0;
-    public int DurationPerLevel = 0;
+    public float DurationBase = 0;
+    public float DurationPerLevel = 0;
 
     [Header("Skill Effect Condition Duration Modifier:")]
     public SkillEffectValueModifier[] DurationModifiers = new SkillEffectValueModifier[0];
 
     public override void ApplyEffect(Character Owner, ItemSkill SourceItemSkill, Character Target)
     {
-        int FinalDuration = DurationBase;
+        float FinalDuration = DurationBase;
 
         FinalDuration += DurationPerLevel * SourceItemSkill.GetSkillLevel();
 
@@ -30,7 +30,7 @@ public class SkillEffectInflictCondition : SkillEffect {
 
     public override void ApplyEffect(Character Owner, ItemSkill SourceItemSkill, Character Target, int FixedLevel)
     {
-        int FinalDuration = DurationBase;
+        float FinalDuration = DurationBase;
 
         FinalDuration += DurationPerLevel * FixedLevel;
